+++ src/screens/MainScreen/MainScreen.tsx
@@ -511,7 +511,7 @@ const MainScreen: React.FC<any> = ({navigation, route}) => {
               mixpanel.track('กดปิดแคมเปญทอง');
               setShowCampaign('none');
             }}>
-            <Image source={icons.x} style={{width: 10, height: 10}} />
+            <Image source={icons.closeBlack} style={{width: 10, height: 10}} />
           </TouchableOpacity>
         </View>